requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
docx2pdf==0.1.8
idna==3.3
lxml==4.8.0
numpy==1.22.3
pandas==1.4.2
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
pywin32==303
requests==2.27.1
six==1.16.0
tqdm==4.64.0
urllib3==1.26.9
